feat: add about page and routes
