+++ class-components/src/components/About/index.tsx
@@ -1,0 +1,1 @@
+export { About } from './About';
