+++ class-components/src/components/About/style.css
@@ -1,0 +1,8 @@
+.about {
+  margin: 200px auto;
+}
+
+.about__link {
+  display: block;
+  color: inherit;
+}
