+++ class-components/src/components/Footer/style.css
@@ -1,0 +1,16 @@
+.footer {
+  border-top: 2px solid var(--dark-color);
+  padding-top: 50px;
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+}
+
+.footer__link {
+  color: inherit;
+  text-decoration: none;
+}
+
+.footer__link--active {
+  color: var(--accent-color);
+}
