+++ class-components/src/components/Home/index.tsx
@@ -1,0 +1,1 @@
+export { Home } from './Home';
